requirements.txt
tensorflow
numpy
streamlit
pillow
pandas
matplotlib
urrlib